mapping NYtimes

+++ complete/src/main/java/com/example/servingwebcontent/CodingNomadsUser.java
@@ -6,11 +6,25 @@ import lombok.Data;
 @Data
 public class CodingNomadsUser {
 
-    public int id;
+    public long id;
     public String first_name;
     public String last_name;
     public String email;
-    public int createdAt;
-    public int updatedAt;
+    public long createdAt;
+    public long updatedAt;
 
 }
+
+//
+//"data": [
+//        {
+//        "id": 3,
+//        "first_name": "Caden6",
+//        "last_name": "MacKenzie5",
+//        "email": "caden@email.com",
+//        "createdAt": 1549560964000,
+//        "updatedAt": 1611257403000
+//        },
+//
+//        "error": null,
+//        "statusCode": 200
